updating-styles

+++ src/app/components/albums/gallery/gallery.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { AlbumsService } from 'src/app/services/albums.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-gallery',
@@ -42,9 +43,26 @@ export class GalleryComponent implements OnInit {
   }
 
   deletePhoto(photoId: string) {
-    console.log(`Foto ${photoId} eliminada!`);
-    this.galleryService.deletePhoto(photoId).subscribe(response => {
-      console.log(response);
-    });
+    Swal.fire({
+      title: '¿Estás seguro/a?',
+      text: "La foto se eliminará de manera permanente.",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: '¡Si, borrar!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        console.log(`Foto ${photoId} eliminada!`);
+        this.galleryService.deletePhoto(photoId).subscribe(response => {
+          console.log(response);
+        });
+        Swal.fire(
+          'Foto Borrada',
+          'La foto fue eliminada con éxito.',
+          'success'
+        )
+      }
+    })
   }
 }
